move class def to avoid circular import
