+++ src/microstructpy/meshing/__init__.py
@@ -1,5 +1,4 @@
 from microstructpy.meshing.polymesh import PolyMesh
-from microstructpy.meshing.rastermesh import RasterMesh
-from microstructpy.meshing.trimesh import TriMesh
+from microstructpy.meshing.trimesh import RasterMesh, TriMesh
 
 __all__ = ['PolyMesh', 'RasterMesh', 'TriMesh']
